+++ packages/world/ts/node/render-solidity/index.ts
@@ -2,4 +2,3 @@ export * from "./renderSystemInterface";
 export * from "./renderWorldInterface";
 export * from "./types";
 export * from "./worldgen";
-export * from "./systemFunctionSignaturesToAbi";
